fix(gh): unset public address to make streaming work
When running on the Github runners the cloud detection logic determines
the external public address behind NAT via STUN and sets it as public
address. This doesn't really work on the runners and we should set the
public address to map to the private address always. However as preseed
support is currently not working we implement a workaround here by
manually changing the configuration and rewriting all service ones.