fix: set public address to local IP
If we don't set the public address in the preseed, the appliance will
try to determine it by sending a STUN request and check what public IP
the machine sits behind. For the Github runners this isn't going to work
as they are not forwarding any ports from their egress address back to
the runner. Therefor we simply set the public address to the local IP we
also configure as listen address. This will make any local streaming use
case work.